added *args, **kwargs to constructor
When creating a wrapper class for this API, I encountered issues passing down a flattened dict of kwargs to the constructor of Uniswap() class due to it not having *args, **kwargs in it's constructors signature. Any extra arguments that were not explicitily in the constructors signature would throw an error, which makes it less friendly for other software systems to wrap it's code with.

Simply adding *args and **kwargs to the constructor fixes this issue and doesn't intrtoduce any negative effects.