added "simulate" kwarg to make_trade
Simulate can now be specified to return an estimate of the gas fees as well as some other useful metadata for the transaction. Does NOT take into account slippage or price impact for a transaction!!

This was mostly needed so I could find the gas fees ahead of time before pushing an order on-chain and allows me to adjust the amount of network fee token buyback prior to it.